Move AppContext to app/
